// src/types/registerTransformerFetch.ts
export type RegisterTransformerFetch = {
  firstname: String;
  lastname: String;
  email: String;
  password: String;
  password_confirmation: String;
  logo: String;
  organization_name: String;
  organization_email: String;
  country: String;
  state_province: String;
  city: String;
  zipcode: String;
  address: String;
  type: String;
};